Feat: add order dispatched and delivered email templates and views
also modified order confirmation email template to display discount information if present. call for the email views from the superuser dashboard when changing the status.
style searchbox in superuser dashboard and removed possibily to delete laptop and desktop tags, which would throw an error when user click on the laptop and desktop button in the navbar