Maint: fix checkout discount and points
in checkout.views.checkout had to insert the discount id the the session before bag context calculate the bag for us.
in checkout.views.place_order and webhook_handler added code to add points to user profile if the user is not using any discounts.
added code to checkout.models and checkout.admin to display the discounted price in the default django admin panel