Maint: fix html validation
Removed the popover import in base.html causing the document to not validate when logged in as superuser. It was useless anyway because the all the flowbite module it's imported via vanilla js including the popover.
Also removed logs and added closing ul tag in main-nav.html causing the document to not validate when logged in as superuser.